Final Assignment

+++ src/main/java/ImageHoster/controller/ImageController.java
@@ -1,5 +1,6 @@
 package ImageHoster.controller;
 
+import ImageHoster.model.Comment;
 import ImageHoster.model.Image;
 import ImageHoster.model.Tag;
 import ImageHoster.model.User;
@@ -26,6 +27,7 @@ public class ImageController {
 
     @Autowired
     private TagService tagService;
+    private Integer imageId;
 
     //This method displays all the images in the user home page after successful login
     @RequestMapping("images")
@@ -45,9 +47,11 @@ public class ImageController {
     //Also now you need to add the tags of an image in the Model type object
     //Here a list of tags is added in the Model type object
     //this list is then sent to 'images/image.html' file and the tags are displayed
-    @RequestMapping("/images/{title}")
-    public String showImage(@PathVariable("title") String title, Model model) {
-        Image image = imageService.getImageByTitle(title);
+  // @RequestMapping("/images/{title}")
+    @RequestMapping("/images/{imageId}/{title}")
+    public String showImage(@PathVariable("imageId") Integer imageId, Model model) {
+      //  Image image = imageService.getImageByTitle(title);
+        Image image = imageService.getImage(imageId);
         model.addAttribute("image", image);
         model.addAttribute("tags", image.getTags());
         return "images/image";
@@ -92,13 +96,21 @@ public class ImageController {
     //The method first needs to convert the list of all the tags to a string containing all the tags separated by a comma and then add this string in a Model type object
     //This string is then displayed by 'edit.html' file as previous tags of an image
     @RequestMapping(value = "/editImage")
-    public String editImage(@RequestParam("imageId") Integer imageId, Model model) {
+    public String editImage(@RequestParam("imageId") Integer imageId, Model model,HttpSession session) {
+        String returnPath;
+        User user = (User) session.getAttribute("loggeduser");
         Image image = imageService.getImage(imageId);
-
+        if(user!=null && image!=null && (user.getId()!=image.getUser().getId())){
+            String error = "Only the owner of the image can edit the image";
+            model.addAttribute("editError", error);
+            returnPath = "images/image";
+        }else{
+            returnPath = "images/edit";
+        }
         String tags = convertTagsToString(image.getTags());
         model.addAttribute("image", image);
         model.addAttribute("tags", tags);
-        return "images/edit";
+        return returnPath;
     }
 
     //This controller method is called when the request pattern is of type 'images/edit' and also the incoming request is of PUT type
@@ -114,7 +126,7 @@ public class ImageController {
     //The method converts the string to a list of all the tags using findOrCreateTags() method and sets the tags attribute of an image as a list of all the tags
     @RequestMapping(value = "/editImage", method = RequestMethod.PUT)
     public String editImageSubmit(@RequestParam("file") MultipartFile file, @RequestParam("imageId") Integer imageId, @RequestParam("tags") String tags, Image updatedImage, HttpSession session) throws IOException {
-
+        User user = (User) session.getAttribute("loggeduser");
         Image image = imageService.getImage(imageId);
         String updatedImageData = convertUploadedFileToBase64(file);
         List<Tag> imageTags = findOrCreateTags(tags);
@@ -126,13 +138,14 @@ public class ImageController {
         }
 
         updatedImage.setId(imageId);
-        User user = (User) session.getAttribute("loggeduser");
+
         updatedImage.setUser(user);
         updatedImage.setTags(imageTags);
         updatedImage.setDate(new Date());
 
         imageService.updateImage(updatedImage);
-        return "redirect:/images/" + updatedImage.getTitle();
+        //return "redirect:/images/" + updatedImage.getTitle();
+        return "redirect:/images/";
     }
 
 
@@ -140,9 +153,17 @@ public class ImageController {
     //The method calls the deleteImage() method in the business logic passing the id of the image to be deleted
     //Looks for a controller method with request mapping of type '/images'
     @RequestMapping(value = "/deleteImage", method = RequestMethod.DELETE)
-    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId) {
-        imageService.deleteImage(imageId);
-        return "redirect:/images";
+    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId,Model model,HttpSession session) {
+        User user = (User) session.getAttribute("loggeduser");
+        Image image = imageService.getImage(imageId);
+        if(user!=null && image!=null && (user.getId()!=image.getUser().getId())){
+            String error = "Only the owner of the image can edit the image";
+            model.addAttribute("deleteError", error);
+            return showImage(imageId,model);
+        }else{
+            imageService.deleteImage(imageId);
+            return "redirect:/images";
+        }
     }
